requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
pydantic-settings
langgraph
google-generativeai
langchain
langchain-google-genai